refactor attributes into differnt modules
