+++ mrfree/core/__init__.py
@@ -1,4 +1,4 @@
-from . import attributes
+from . import geometry
 from . import region
 from . import tract
-__all__ = ['attributes', 'region', 'tract']
+__all__ = ['geometry.py', 'region', 'tract']
